abstracted generateStateMachine code
