+++ workflow/serverless.yml
@@ -34,8 +34,8 @@ functions:
         method: get
         cors: true
 
-  addWorkflowToProject:
-    handler: project/addWorkflowToProject.handler
+  addWorkflow:
+    handler: workflow/addWorkflow.handler
     role: ${env:LAMBDA_ROLE}
     events:
       - http:
